+++ Projects/App/Sources/LivithApp.swift
@@ -1,20 +1,16 @@
 import SwiftUI
-import UserNotifications
 
-import KakaoSDKAuth
 import KakaoSDKCommon
 import LivithFoundation
 
 @main
 struct LivithApp: App {
     @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
-    @State private var isLaunchScreenVisible = true
 
     // MARK: - LifeCycle
 
     init() {
         registerDependency()
-
         initializeKakaoSDK()
     }
 
@@ -22,46 +18,11 @@ struct LivithApp: App {
         WindowGroup {
             AppRootView()
                 .task {
-                    await requestNotificationAuthorization()
+                    await NotificationService.shared.requestAuthorization()
                 }
                 .onOpenURL { url in
-                    handleOpenURL(url)
+                    DeepLinkService.shared.handle(url: url)
                 }
-        }
-    }
-}
-
-// MARK: - URL Handling
-
-private extension LivithApp {
-    func handleOpenURL(_ url: URL) {
-        if AuthApi.isKakaoTalkLoginUrl(url) {
-            _ = AuthController.handleOpenUrl(url: url)
-            return
-        }
-
-        if url.scheme == "livith" {
-            handleDeepLink(url)
-        }
-    }
-
-    func handleDeepLink(_ url: URL) {
-        guard let host = url.host else { return }
-
-        switch host {
-        case "concert":
-            if let concertIDString = url.pathComponents.dropFirst().first,
-               let concertID = Int(concertIDString) {
-                NotificationCenter.default.post(
-                    name: .openConcertDetail,
-                    object: nil,
-                    userInfo: ["concertID": concertID]
-                )
-            }
-        case "home":
-            break
-        default:
-            break
         }
     }
 }
@@ -73,19 +34,4 @@ private extension LivithApp {
         guard let kakaoAppKey = Bundle.main.infoDictionary?["NATIVE_APP_KEY"] as? String else { return }
         KakaoSDK.initSDK(appKey: kakaoAppKey)
     }
-}
-
-// MARK: - Notification
-
-private extension LivithApp {
-    func requestNotificationAuthorization() async {
-        let options: UNAuthorizationOptions = [.alert, .badge, .sound]
-        _ = try? await UNUserNotificationCenter.current().requestAuthorization(options: options)
-    }
-}
-
-// MARK: - Notification.Name
-
-public extension Notification.Name {
-    static let openConcertDetail = Notification.Name("openConcertDetail")
 }
